jekyllandhyde/requirements.txt
fastapi
uvicorn
openai
numpy
pypdf
python-dotenv
